+++ Clients/GogoAnimeClient.py
@@ -43,6 +43,9 @@ class GogoAnimeClient(BaseClient):
         meta = {}
         soup = self._get_bsoup(link)
         # self.logger.debug(f'bsoup response for {link = }: {soup}')
+        if soup is None:
+            return None
+
         for detail in soup.select(self.series_info_element):
             line = detail.text.strip()
             if ':' in line:
@@ -93,14 +96,16 @@ class GogoAnimeClient(BaseClient):
             link = element['href']
             if link.startswith('/'):
                 link = self.base_url + link
-            item = {'title': title, 'link': link}
-            # get every search result details
-            item.update(self._get_series_info(link))
-            item['year'] = item['Released']
-            # add index to every search result
-            search_results[idx] = item
-            self._show_search_results(idx, item)
-            idx += 1
+            data = self._get_series_info(link)
+            if data is not None:
+                item = {'title': title, 'link': link}
+                # get every search result details
+                item.update(data)
+                item['year'] = item['Released']
+                # add index to every search result
+                search_results[idx] = item
+                self._show_search_results(idx, item)
+                idx += 1
 
         return search_results