+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
  - Version 2.12.3 [2024-06-15]
-   - Added recursive input for user prompts to avoid hassle of reloading from the start.
+   - Added recursive input for user prompts and whole UDB to avoid hassle of reloading from the start.
    - Updated TMDB Client: Removed dependency on vidsrc while displaying search results.
    - Updated HLS Downloader: Skip non-downloadable subtitles for Movies/Series.
    - Fixed HLS Downloader: Unable to download series with special characters like '#'.